Update vcfCheck.R
Duplicate checks have been removed, and numeric encoding of the VCF genotypes has been added, storing the result in the metadata (geno_coded). This is done within this function because the same procedure was repeated in several downstream functions, which was inefficient. By encoding the genotypes once here, we can reuse this information throughout the workflow without recalculating it.